+++ src/main/java/com/udemy/course/resources/UserResource.java
@@ -24,7 +24,7 @@ public class UserResource {
 
 	@GetMapping
 	public ResponseEntity<User> findAll() {
-		User user = new User(1l, "Maria", "maria@gmail.com", "99999999", "12345");
+		User user = new User((int) 1l, "Maria", "maria@gmail.com", "99999999", "12345");
 		return ResponseEntity.ok().body(user);
 	}
 }
